requirements.txt
flask
google-cloud-translate
python-dotenv
